feat: add weather page with climate information display
- Introduced Climate.vue component for displaying weather data, including temperature, humidity, wind speed, and moon phase.
- Created climate.js for managing weather data retrieval from an external API.
- Added climate.css for styling the weather page with a responsive design and dark mode adjustments.
- Implemented a loading skeleton for improved user experience during data fetching.
