+++ src/pages/climate/climate.css
@@ -1,0 +1,110 @@
+.weather-page {
+  min-height: 100vh;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+.bg-image {
+  background-image: url('https://images.unsplash.com/photo-1590077428593-a33c3fb61d8d?q=80&w=2096');
+  background-size: cover;
+  background-position: center;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: -2;
+}
+
+.gradient-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(135deg, rgba(0,0,0,0.3) 0%, rgba(0,0,0,0.5) 100%);
+  z-index: -1;
+}
+
+.content-container {
+  width: 100%;
+  max-width: 500px;
+  margin: 0 auto;
+  z-index: 1;
+}
+
+.search-container {
+  width: 100%;
+  margin-top: 20px;
+}
+
+.glass-input {
+  background: rgba(255, 255, 255, 0.1) !important;
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 12px;
+  color: white;
+}
+
+.glass-input .q-field__control {
+  background: transparent !important;
+}
+
+.glass-input .q-field__native,
+.glass-input .q-field__prefix,
+.glass-input .q-field__suffix,
+.glass-input .q-icon {
+  color: white;
+}
+
+.glass-card {
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  border-radius: 16px;
+  color: white;
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  transition: all 0.3s ease;
+}
+
+.detail-card {
+  background: rgba(255, 255, 255, 0.1);
+  border-radius: 12px;
+  transition: all 0.3s ease;
+}
+
+.detail-card:hover {
+  background: rgba(255, 255, 255, 0.2);
+}
+
+.weather-icon {
+  width: 80px;
+  height: 80px;
+  filter: drop-shadow(0 0 8px rgba(255, 255, 255, 0.5));
+}
+
+.moon-icon {
+  width: 60px;
+  height: 60px;
+  filter: drop-shadow(0 0 5px rgba(255, 255, 255, 0.3));
+}
+
+.location-container {
+  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+/* Dark mode adjustments */
+.body--dark .glass-card,
+.body--dark .detail-card,
+.body--dark .glass-input {
+  background: rgba(0, 0, 0, 0.2);
+  border-color: rgba(255, 255, 255, 0.1);
+}
+
+.body--dark .detail-card:hover {
+  background: rgba(0, 0, 0, 0.3);
+}
